% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{check_inherits}
\alias{check_inherits}
\title{Test class of object}
\usage{
check_inherits(x, cl, allow_null = TRUE, xname = deparse(substitute(x)))
}
\arguments{
\item{x}{Object to check.}

\item{cl}{Character: class to check against.}

\item{allow_null}{Logical: If TRUE, NULL values are allowed and return early.}
}
\value{
Called for side effects. Throws an error if checks fail.
}
\description{
Test class of object
}
\details{
Exported as internal function for use by other rtemis packages.
}
\author{
EDG
}
\keyword{internal}
