% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_checks.R
\name{clean_int}
\alias{clean_int}
\title{Clean integer input}
\usage{
clean_int(x, xname = deparse(substitute(x)))
}
\arguments{
\item{x}{Double or integer vector to check.}

\item{xname}{Character: Name of x, for error messages.}
}
\value{
Integer vector
}
\description{
Clean integer input
}
\details{
The goal is to return an integer vector.
If the input is integer, it is returned as is.
If the input is numeric, it is coerced to integer only if the numeric values are integers,
otherwise an error is thrown.
}
\examples{
\dontrun{
clean_int(6L)
clean_int(3)
clean_int(12.1) # Error
clean_int(c(3, 5, 7))
clean_int(c(3, 5, 7.01)) # Error
}
}
\author{
EDG
}
