% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{df_movecolumn}
\alias{df_movecolumn}
\title{Move data frame column}
\usage{
df_movecolumn(x, colname, to = ncol(x))
}
\arguments{
\item{x}{data.frame.}

\item{colname}{Character: Name of column you want to move.}

\item{to}{Integer: Which column position to move the vector to.
Default = \code{ncol(x)} i.e. the last column.}
}
\value{
data.frame
}
\description{
Move data frame column
}
\examples{
\dontrun{
ir <- df_movecolumn(iris, colname = "Species", to = 1L)
}
}
\author{
EDG
}
