% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{dt_names_by_attr}
\alias{dt_names_by_attr}
\title{List column names by attribute}
\usage{
dt_names_by_attr(x, attribute, exact = TRUE, sorted = TRUE)
}
\arguments{
\item{x}{data.table: Input data.table.}

\item{attribute}{Character: name of attribute.}

\item{exact}{Logical: If TRUE, use exact matching.}

\item{sorted}{Logical: If TRUE, sort the output.}
}
\value{
Character vector.
}
\description{
List column names by attribute
}
\examples{
library(data.table)
x <- data.table(
  id = 1:5,
  sbp = rnorm(5, 120, 15),
  dbp = rnorm(5, 80, 10),
  paO2 = rnorm(5, 90, 10),
  paCO2 = rnorm(5, 40, 5)
)
setattr(x[["id"]], "source", "demographics")
setattr(x[["sbp"]], "source", "outpatient")
setattr(x[["dbp"]], "source", "outpatient")
setattr(x[["paO2"]], "source", "icu")
setattr(x[["paCO2"]], "source", "icu")

dt_names_by_attr(x, "source", "outpatient")
}
\author{
EDG
}
