% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{dt_pctmissing}
\alias{dt_pctmissing}
\title{Get percent of missing values from every column}
\usage{
dt_pctmissing(x, verbosity = 1L)
}
\arguments{
\item{x}{data.frame or data.table}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
list
}
\description{
Get percent of missing values from every column
}
\examples{
library(data.table)
x <- data.table(a = c(1, 2, NA, 4), b = c(NA, NA, 3, 4), c = c("A", "B", "C", NA))
dt_pctmissing(x)
}
\author{
EDG
}
