% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.table.R
\name{dt_set_autotypes}
\alias{dt_set_autotypes}
\title{Set column types automatically}
\usage{
dt_set_autotypes(x, cols = NULL, verbosity = 1L)
}
\arguments{
\item{x}{data.table: Input data.table. Will be modified \emph{\strong{in-place}}, if needed.}

\item{cols}{Character vector: columns to work on. If not defined, will work on all
columns}

\item{verbosity}{Integer: Verbosity level.}
}
\value{
data.table, invisibly.
}
\description{
This function inspects a data.table and attempts to identify columns that should be
numeric but have been read in as character, and fixes their type \emph{\strong{in-place}}.
This can happen when one or more fields contain non-numeric characters, for example.
}
\examples{
library(data.table)
x <- data.table(
  id = 8001:8006,
  a = c("3", "5", "undefined", "21", "4", NA),
  b = c("mango", "banana", "tangerine", NA, "apple", "kiwi"),
  c = c(1, 2, 3, 4, 5, 6)
)
str(x)
# ***in-place*** operation means no assignment is needed
dt_set_autotypes(x)
str(x)

# Try excluding column 'a' from autotyping
x <- data.table(
  id = 8001:8006,
  a = c("3", "5", "undefined", "21", "4", NA),
  b = c("mango", "banana", "tangerine", NA, "apple", "kiwi"),
  c = c(1, 2, 3, 4, 5, 6)
)
str(x)
# exclude column 'a' from autotyping
dt_set_autotypes(x, cols = setdiff(names(x), "a"))
str(x)
}
\author{
EDG
}
