% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{fmt_gradient}
\alias{fmt_gradient}
\title{Gradient text}
\usage{
fmt_gradient(x, colors, bold = FALSE, output_type = c("ansi", "html", "plain"))
}
\arguments{
\item{x}{Character: Text to colorize.}

\item{colors}{Character vector: Colors to use for the gradient.}

\item{bold}{Logical: If TRUE, make text bold.}

\item{output_type}{Character: Output type ("ansi", "html", "plain").}
}
\value{
Character: Text with gradient color applied.
}
\description{
Gradient text
}
\author{
EDG
}
