% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{get_output_type}
\alias{get_output_type}
\title{Get output type}
\usage{
get_output_type(output_type = c("ansi", "html", "plain"), filename = NULL)
}
\arguments{
\item{output_type}{Character vector of output types.}

\item{filename}{Optional Character: Filename for output.}
}
\value{
Character with selected output type.
}
\description{
Get output type for printing text.
}
\details{
Exported as internal function for use by other rtemis packages.
}
\examples{
get_output_type()
}
\author{
EDG
}
