% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_df.R
\name{getnames}
\alias{getnames}
\alias{getfactornames}
\alias{getnumericnames}
\alias{getlogicalnames}
\alias{getcharacternames}
\alias{getdatenames}
\title{Get names by string matching or class}
\usage{
getnames(
  x,
  pattern = NULL,
  starts_with = NULL,
  ends_with = NULL,
  ignore_case = TRUE
)

getfactornames(x)

getnumericnames(x)

getlogicalnames(x)

getcharacternames(x)

getdatenames(x)
}
\arguments{
\item{x}{object with \code{names()} method.}

\item{pattern}{Character: pattern to match anywhere in names of x.}

\item{starts_with}{Character: pattern to match in the beginning of names of x.}

\item{ends_with}{Character: pattern to match at the end of names of x.}

\item{ignore_case}{Logical: If TRUE, well, ignore case.}
}
\value{
Character vector of matched names.
}
\description{
Get names by string matching or class
}
\details{
For \code{getnames()} only:
\code{pattern}, \code{starts_with}, and \code{ends_with} are applied sequentially.
If more than one is provided, the result will be the intersection of all matches.
}
\examples{
getnames(iris, starts_with = "Sepal")
getnames(iris, ends_with = "Width")
}
\author{
EDG
}
