% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{gray}
\alias{gray}
\title{Gray text}
\usage{
gray(x, output_type = c("ansi", "html", "plain"))
}
\arguments{
\item{x}{Character: Text to format}

\item{output_type}{Character: Output type ("ansi", "html", "plain")}
}
\value{
Character: Formatted text with gray styling
}
\description{
A \code{fmt()} convenience wrapper for making text gray.
}
\details{
Can be useful in contexts where muted is not supported.
}
\author{
EDG
}
