% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{green}
\alias{green}
\title{Make text green}
\usage{
green(..., bold = FALSE)
}
\arguments{
\item{...}{Character: Text to colorize.}

\item{bold}{Logical: If TRUE, make text bold.}
}
\description{
Make text green
}
\author{
EDG
}
