% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{highlight}
\alias{highlight}
\title{Highlight text}
\usage{
highlight(x, pad = 0L, output_type = c("ansi", "html", "plain"))
}
\arguments{
\item{x}{Character: Text to highlight.}

\item{pad}{Integer: Number of spaces to pad before text.}

\item{output_type}{Character: Output type ("ansi", "html", "plain").}
}
\value{
Character: Formatted text with highlight.
}
\description{
A \code{fmt()} convenience wrapper for highlighting text.
}
\author{
EDG
}
\keyword{internal}
