% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_data.R
\name{index_col_by_attr}
\alias{index_col_by_attr}
\title{Index columns by attribute name & value}
\usage{
index_col_by_attr(x, name, value)
}
\arguments{
\item{x}{data.frame or similar.}

\item{name}{Character: Name of attribute.}

\item{value}{Character: Value of attribute.}
}
\value{
Integer vector.
}
\description{
Index columns by attribute name & value
}
\examples{
library(data.table)
x <- data.table(
  id = 1:5,
  sbp = rnorm(5, 120, 15),
  dbp = rnorm(5, 80, 10),
  paO2 = rnorm(5, 90, 10),
  paCO2 = rnorm(5, 40, 5)
)
setattr(x[["sbp"]], "source", "outpatient")
setattr(x[["dbp"]], "source", "outpatient")
setattr(x[["paO2"]], "source", "icu")
setattr(x[["paCO2"]], "source", "icu")
index_col_by_attr(x, "source", "icu")
}
\author{
EDG
}
