% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{italic}
\alias{italic}
\title{Make text italic}
\usage{
italic(text, output_type = c("ansi", "html", "plain"))
}
\arguments{
\item{text}{Character: Text to make italic}

\item{output_type}{Character: Output type ("ansi", "html", "plain")}
}
\value{
Character: Formatted text with italic styling
}
\description{
A \code{fmt()} convenience wrapper for making text italic.
}
\author{
EDG
}
