% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{muted}
\alias{muted}
\title{Muted text}
\usage{
muted(x, output_type = c("ansi", "html", "plain"))
}
\arguments{
\item{x}{Character: Text to format}

\item{output_type}{Character: Output type ("ansi", "html", "plain")}
}
\value{
Character: Formatted text with muted styling
}
\description{
A \code{fmt()} convenience wrapper for making text muted.
}
\author{
EDG
}
