% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{orange}
\alias{orange}
\title{Make text orange}
\usage{
orange(..., bold = FALSE)
}
\arguments{
\item{...}{Character: Text to colorize.}

\item{bold}{Logical: If TRUE, make text bold.}
}
\description{
Make text orange
}
\author{
EDG
}
