% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msg.R
\name{pcat}
\alias{pcat}
\title{Pad-cat}
\usage{
pcat(left, right, pad = 17, newline = TRUE)
}
\arguments{
\item{left}{Character: Left string to pad and print.}

\item{right}{Character: Right string to print after left.}

\item{pad}{Integer: Total width to pad the left string to.}

\item{newline}{Logical: If TRUE, print a newline after the right string.}
}
\description{
Pad and concatenate two strings, with optional newline.
}
\examples{
\dontrun{
{
  msg("Hello")
  pcat("super", "wow")
  pcat(NULL, "oooo")
}
}
}
\author{
EDG
}
