% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_art.R
\name{pkglogo}
\alias{pkglogo}
\title{pkglogo}
\usage{
pkglogo(
  pkg,
  filename = NULL,
  fmt_fn = color_txt_columns,
  args = list(color_left = rtemis_colors[["red"]], color_right =
    rtemis_colors[["orange"]], output_type = "ansi"),
  pad = 0L
)
}
\arguments{
\item{pkg}{Character: Package name.}

\item{filename}{Character: Filename of the logo file (without path). If NULL,
defaults to "\if{html}{\out{<pkg>}}.utf8".}

\item{fmt_fn}{Function: "color_txt_columns", "color_text_rows" Formatting function to apply to
the logo text.}

\item{args}{List: Arguments to pass to \code{fmt_fn}.}

\item{pad}{Integer: Left-pad output with this many spaces.}
}
\value{
Character: Formatted logo text.
}
\description{
Display a colored ASCII logo for a package from its inst/resources directory.
}
\examples{
\dontrun{
# In .onAttach() function:
.onAttach <- function(libname, pkgname) {
  packageStartupMessage(pkglogo(pkg = pkgname))
}
}
}
\author{
EDG
}
