% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_strings.R
\name{plain}
\alias{plain}
\title{Force plain text when using \code{message()}}
\usage{
plain(x)
}
\arguments{
\item{x}{Character: Text to be output to console.}
}
\value{
Character: Text with ANSI escape codes removed.
}
\description{
Force plain text when using \code{message()}
}
\author{
EDG
}
