% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_print.R
\name{printls}
\alias{printls}
\title{Pretty print list}
\usage{
printls(
  x,
  prefix = "",
  pad = 2L,
  item_format = bold,
  maxlength = 4L,
  center_title = TRUE,
  title = NULL,
  title_newline = TRUE,
  newline_pre = FALSE,
  format_fn_rhs = ddSci,
  print_class = TRUE,
  abbrev_class_n = 3L,
  print_df = FALSE,
  print_S4 = FALSE,
  limit = 12L
)
}
\arguments{
\item{x}{list or object that will be converted to a list.}

\item{prefix}{Character: Optional prefix for names.}

\item{pad}{Integer: Pad output with this many spaces.}

\item{item_format}{Formatting function for list item names.}

\item{maxlength}{Integer: Maximum length of items to show using \code{headdot()} before truncating with ellipsis.}

\item{center_title}{Logical: If TRUE, autopad title for centering, if present.}

\item{title}{Character: Optional title to print before list.}

\item{title_newline}{Logical: If TRUE, print title on new line.}

\item{newline_pre}{Logical: If TRUE, print newline before list.}

\item{format_fn_rhs}{Formatting function for right-hand side values.}

\item{print_class}{Logical: If TRUE, print abbreviated class of object.}

\item{abbrev_class_n}{Integer: Number of characters to abbreviate class names to.}

\item{print_df}{Logical: If TRUE, print data frame contents, otherwise print n rows and columns.}

\item{print_S4}{Logical: If TRUE, print S4 object contents, otherwise print class name.}

\item{limit}{Integer: Maximum number of items to show. Use -1 for unlimited.}
}
\description{
Pretty print a list (or data frame) recursively
}
\details{
Data frames in R began life as lists
}
\author{
EDG
}
\keyword{internal}
