% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/1_S7_init.R
\name{repr}
\alias{repr}
\title{String representation}
\usage{
repr(x, ...)
}
\arguments{
\item{x}{Object to represent as a string.}

\item{...}{Additional arguments passed to methods.}
}
\value{
Character string representation of the object.
}
\description{
String representation
}
\author{
EDG
}
