% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rt_letters}
\alias{rt_letters}
\title{Construct an n-length vector of letters}
\usage{
rt_letters(n = 100, caps = FALSE)
}
\arguments{
\item{n}{Length of vector to return}

\item{caps}{Logical: If TRUE, return all caps}
}
\description{
Returns an n-length vector of the latin alphabet, replicating for every 26 characters
}
