% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemis_color_system.R
\docType{data}
\name{rtemis_colors}
\alias{rtemis_colors}
\title{rtemis Color System}
\format{
A named list with the following elements:
\describe{
\item{red}{"kaimana red"}
\item{blue}{"kaimana light blue"}
\item{green}{"kaimana medium green"}
\item{orange}{"coastside orange"}
\item{teal}{"rtemis teal"}
\item{purple}{"rtemis purple"}
\item{magenta}{"rtemis magenta"}
\item{highlight_col}{"highlight color"}
\item{object}{"rtemis teal"}
\item{info}{"lmd burgundy"}
\item{outer}{"kaimana red"}
\item{tuner}{"coastside orange"}
}
}
\usage{
rtemis_colors
}
\description{
A named list of colors used consistently across all packages
in the rtemis ecosystem.
}
\details{
Colors are provided as hex strings.
}
\examples{
rtemis_colors[["orange"]]

rtemis_colors[["teal"]]
}
\author{
EDG
}
\keyword{datasets}
