% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemisutils-package.R
\docType{package}
\name{rtemisutils-package}
\alias{rtemis.utils}
\alias{rtemis.utils-package}
\alias{rtemisutils-package}
\title{\pkg{rtemisutils}: Rtemis Utilities}
\description{
Utilities for Rtemis R Packages
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.rtemis.org}
  \item Report bugs at \url{https://github.com/rtemis-org/rtemis.utils/issues}
}

}
\author{
\strong{Maintainer}: E.D. Gennatas \email{gennatas@gmail.com} (\href{https://orcid.org/0000-0001-9280-3609}{ORCID}) [copyright holder]

}
