% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{size}
\alias{size}
\title{Size of object}
\usage{
size(x, verbosity = 1L)
}
\arguments{
\item{x}{any object with \code{length()} or \code{dim()}.}

\item{verbosity}{Integer: Verbosity level. If > 0, print size to console}
}
\value{
Integer vector with length equal to the number of dimensions of \code{x}, invisibly.
}
\description{
Returns the size of an object
}
\details{
If \code{dim(x)} is NULL, returns \code{length(x)}.
}
\examples{
x <- rnorm(20)
size(x)
# 20
x <- matrix(rnorm(100), 20, 5)
size(x)
# 20  5
}
\author{
EDG
}
