% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt.R
\name{thin}
\alias{thin}
\title{Make text thin/light}
\usage{
thin(text, output_type = c("ansi", "html", "plain"))
}
\arguments{
\item{text}{Character: Text to make thin}

\item{output_type}{Character: Output type ("ansi", "html", "plain")}
}
\value{
Character: Formatted text with thin/light styling
}
\description{
A \code{fmt()} convenience wrapper for making text thin/light.
}
\author{
EDG
}
