% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtemis_color_system.R
\docType{data}
\name{rtemis_colors}
\alias{rtemis_colors}
\title{rtemis Color System}
\format{
A named list with the following elements:
\describe{
\item{rt_red}{"kaimana red"}
\item{rt_blue}{"kaimana light blue"}
\item{rt_green}{"kaimana medium green"}
\item{rt_orange}{"coastside orange"}
\item{rt_teal}{"rtemis teal"}
\item{rt_purple}{"rtemis purple"}
\item{rt_magenta}{"rtemis magenta"}
\item{highlight_col}{"rtemis teal"}
\item{col_object}{"rtemis teal"}
\item{col_info}{"lmd burgundy"}
\item{col_outer}{"kaimana red"}
\item{col_tuner}{"coastside orange"}
}
}
\usage{
rtemis_colors
}
\description{
A named list of colors used consistently across all packages
in the rtemis ecosystem.
}
\details{
Colors are provided as hex strings.
}
\examples{
rtemis_colors[["rt_teal"]]

rtemis_colors[["teal"]]
}
\author{
EDG
}
\keyword{datasets}
